First Commit
I made this project quite a while ago, back then I didn't know how to use tables or dictionaries, so the code to get a certain price is really messy and bad.

Expect this to be updated in the near future. I will probably read the price data from a database in a certain type of file.